command.cpp - Signed/unsigned fixes daInitLayer - began rewriting class to use damatrix rotation functions. To be completed tomorrow. daPf - damatrix-derived phase functions. Automatically covers rotation in case of P(alpha). Needs completion to sort between P(alpha) and P(valmap). debug_mem - Namespace change in function. Unimportant. phaseFunc - made pfRotator obsolete. Code that calls it will throw. error - xError now has xObsolete and xDefective. Indicates obsolete or known defective code as throw. Called by base classes defective / obsolete. Makes marking whole class easy!
git-svn-id: file:///mnt/svn/rtmath@181 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath/command.cpp
@@ -36,7 +36,7 @@ namespace rtmath {
 				} while (end != endpos);
 
 				// Post-parsing (search for empty commands, remove whitespace, ...)
-				for (int i=0;i<rawCommand.size();i++)
+				for (int i=0;i<(int)rawCommand.size();i++)
 				{
 					if (rawCommand[i].size() == 0) continue;
 					command.push_back(rawCommand[i]);
@@ -46,7 +46,7 @@ namespace rtmath {
 			// Now to move on to actual command processing:
 			// Loop through all commands and execute them.
 			// Catch any errors returned and dispatch them to the appropriate listener/
-			for (int i=0; i<command.size(); i++)
+			for (int i=0; i<(int)command.size(); i++)
 			{
 
 			}
